refactor to include typescript.types exported
