+++ src/JsonSchemaEditor/JsonSchemaEditor.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import { useState } from "@hookstate/core";
 import { useSchemaState, defaultSchema } from "./state";
-import { SchemaEditorProps } from "./JsonSchemaEditor.types";
+import { SchemaEditorProps } from "../JsonSchemaEditor.types";
 import { Flex, ChakraProvider, theme } from "@chakra-ui/react";
 
 import { SchemaRoot } from "./schema-root";
@@ -9,49 +9,49 @@ import { Whoops } from "./whoops";
 import { SchemaObject } from "./schema-object";
 import { SchemaArray } from "./schema-array";
 
-export * from "./JsonSchemaEditor.types";
+export * from "../JsonSchemaEditor.types";
 
 export const JsonSchemaEditor = (props: SchemaEditorProps) => {
-  const { onSchemaChange, readOnly, data } = props;
+	const { onSchemaChange, readOnly, data } = props;
 
-  const schemaState = useSchemaState({
-    jsonSchema: data ?? defaultSchema(),
-    isReadOnly: readOnly ?? false,
-    fieldId: 0,
-  });
+	const schemaState = useSchemaState({
+		jsonSchema: data ?? defaultSchema(),
+		isReadOnly: readOnly ?? false,
+		fieldId: 0,
+	});
 
-  const jsonSchemaState = useState(schemaState.jsonSchema);
+	const jsonSchemaState = useState(schemaState.jsonSchema);
 
-  return (
-    <ChakraProvider theme={theme}>
-      {schemaState.isValidSchema ? (
-        <Flex m={2} direction="column">
-          <SchemaRoot
-            onSchemaChange={onSchemaChange}
-            schemaState={schemaState.jsonSchema}
-            isReadOnly={schemaState.isReadOnly}
-          />
+	return (
+		<ChakraProvider theme={theme}>
+			{schemaState.isValidSchema ? (
+				<Flex m={2} direction="column">
+					<SchemaRoot
+						onSchemaChange={onSchemaChange}
+						schemaState={schemaState.jsonSchema}
+						isReadOnly={schemaState.isReadOnly}
+					/>
 
-          {jsonSchemaState.type.value === "object" && (
-            <SchemaObject
-              schemaState={jsonSchemaState}
-              isReadOnly={schemaState.isReadOnly ?? false}
-            />
-          )}
+					{jsonSchemaState.type.value === "object" && (
+						<SchemaObject
+							schemaState={jsonSchemaState}
+							isReadOnly={schemaState.isReadOnly ?? false}
+						/>
+					)}
 
-          {jsonSchemaState.type.value === "array" && (
-            <SchemaArray
-              schemaState={jsonSchemaState}
-              isReadOnly={schemaState.isReadOnly ?? false}
-            />
-          )}
-        </Flex>
-      ) : (
-        <Flex alignContent="center" justifyContent="center">
-          <Whoops />
-        </Flex>
-      )}
-      {/* <Modal
+					{jsonSchemaState.type.value === "array" && (
+						<SchemaArray
+							schemaState={jsonSchemaState}
+							isReadOnly={schemaState.isReadOnly ?? false}
+						/>
+					)}
+				</Flex>
+			) : (
+				<Flex alignContent="center" justifyContent="center">
+					<Whoops />
+				</Flex>
+			)}
+			{/* <Modal
         isOpen={localState.isAdvancedOpen.get()}
         finalFocusRef={focusRef}
         size="lg"
@@ -77,6 +77,6 @@ export const JsonSchemaEditor = (props: SchemaEditorProps) => {
           </ModalFooter>
         </ModalContent>
       </Modal> */}
-    </ChakraProvider>
-  );
+		</ChakraProvider>
+	);
 };
