--- src/JsonSchemaEditor/index.ts
@@ -1,1 +1,0 @@
-export { JsonSchemaEditor as default } from "./JsonSchemaEditor";
