+++ src/index.ts
@@ -1,0 +1,1 @@
+export { JsonSchemaEditor as default } from "./JsonSchemaEditor/JsonSchemaEditor";
